Move opener to lib
